readme: change gbarunner3 frontend version
The twilight menu version of the frontend
expects GBAR3 to be in '/_nds/TWiLightMenu/emulators'.
However, the setup instructions tell the user to place
GBAR3 on the root of their SD card, which means the frontend
will not be able to find it. Swapping out the twilight
frontend with the normal frontend will fix this issue.